backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.3
openai==2.1.0
packaging==25.0
playwright==1.55.0
psycopg2==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.9.0
pyee==13.0.0
python-dotenv==1.1.1
requests==2.32.5
schedule==1.2.2
sniffio==1.3.1
SQLAlchemy==2.0.43
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
